Finish moving code out of models/__init__.py
