+++ auto_causality/models/dummy.py
@@ -3,7 +3,7 @@ from typing import List
 import numpy as np
 import pandas as pd
 
-from auto_causality.models import DoWhyMethods
+from auto_causality.models.wrapper import DoWhyMethods, DoWhyWrapper
 from auto_causality.scoring import ate
 
 
@@ -29,3 +29,8 @@ class DummyModel(DoWhyMethods):
     def predict(self, X: pd.DataFrame) -> np.ndarray:
         mean_, _, _ = ate(X[self.treatment], X[self.outcome])
         return np.ones(len(X)) * mean_ * (1 + 0.01 * np.random.normal(size=(len(X),)))
+
+
+class Dummy(DoWhyWrapper):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, inner_class=DummyModel, **kwargs)
